Dat(a)Con(tainer), Patron stuff, fixes and adjustments - Sound is now muted when skipping ahead in the sound queue, to avoid bursting eardrums - more items for the upcoming things - Patron stuff for setup, including the application of stats in the beginning and the init() call in the PlayerCreator - Added DataContainer for ItemInstance, known as Datcon. It allows you to store any datatype within an ItemInstance for use in the original Item class (see: CentrifugeItem.java)

+++ src/net/vami/game/interactable/entity/WolfEntity.java
@@ -30,7 +30,7 @@ public class WolfEntity extends Entity {
     @Override
     public void hurt(Interactable source, float amount, DamageType damageType) {
         super.hurt(source, amount, damageType);
-        if (this.getHealth() < (float) this.getMaxHealth() / 2) {
+        if (this.getHealth() < this.getMaxHealth() / 2) {
             addTask(Tasks.ABILITY_OR_TARGET, 10);
             removeTask(Tasks.TARGET_ANY_AND_ATTACK);
             addTask(Tasks.TARGET_AND_ATTACK, 10);
